// frontend/src/generated/types/clientUserRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface ClientUserRequest {
  sysId?: string;
  userId?: string;
  userNm?: string;
  userRole?: string;
  plantCd?: string;
  plantNm?: string;
}